Design fixes and code improvements
