+++ app/javascript/packs/components/RestaurantsList.jsx
@@ -1,44 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-
-class RestaurantRow extends React.Component {
-  static calcStars(rating) {
-    let stars = '';
-    for (let i = 0; i < rating; i = i + 1) {
-      stars = stars + '★';
-    }
-    return stars;
-  }
-
-  render() {
-    const restaurant = this.props.restaurant;
-    const imagePath = 'images/' + restaurant.cuisine_icon;
-    return (
-        <div className="restaurant_cell">
-          <div className="cuisine_icon_div">
-              <img className="cuisine_icon" src={imagePath} alt={restaurant.cuisine_icon} />
-          </div>
-          <div className="restaurant_main_info">
-              {restaurant.name}<br/><br/>
-              Delivery Time: {restaurant.max_delivery_time} Minutes<br/>
-              {restaurant.address}<br/>
-          </div>
-          <div className="restaurant_secondary_info">
-            <div className="ten_bis_icon">
-                <img className={restaurant.accepts_ten_bis ? 'accepts_ten_bis' : 'not_accepts_ten_bis'} />
-            </div>
-            <div className="restaurant_rating">
-                {RestaurantRow.calcStars(restaurant.rating)}
-            </div>
-          </div>
-        </div>
-    );
-  }
-}
-
-RestaurantRow.propTypes = {
-  restaurant: PropTypes.object.isRequired,
-};
+import {RestaurantCell, EmptyRestaurantCell} from './RestaurantCell';
 
 class FilterableRestaurantTable extends React.Component {
   render() {
@@ -89,12 +51,11 @@ class FilterableRestaurantTable extends React.Component {
     }
     filtersArray.map(applyFilter);
 
-
     if (restaurantList && restaurantList.length > 0) {
       const rows = [];
       restaurantList.forEach((restaurant) => {
         rows.push(
-          <RestaurantRow
+          <RestaurantCell
             restaurant = {restaurant}
             key = {restaurant.id}
           />
@@ -105,7 +66,7 @@ class FilterableRestaurantTable extends React.Component {
     }
     // error handling
     return (
-      <h2>No restaurants found</h2>
+      <EmptyRestaurantCell />
     );
   }
 }
