+++ app/views/layouts/index.html.erb
@@ -3,10 +3,6 @@
 <head>
   <title>WeEat</title>
   <%= csrf_meta_tags %>
-
-  <div id="container">
-    <!-- This element's contents will be replaced with your component. -->
-  </div>
 </head>
 
 <body>
